Move cursor impls to their own file
